refactor: add sass, change structure
