+++ management-app/package.json
@@ -27,6 +27,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.1",
     "react-scripts": "5.0.1",
+    "sass": "^1.56.1",
     "sass-loader": "^13.1.0",
     "typescript": "^4.8.4",
     "web-vitals": "^2.1.4"
